refactor: add index files and routes to pages
