+++ qr-code/node/web/frontend/hooks/useShopifyQuery.js
@@ -1,7 +1,7 @@
 import { useQuery } from 'react-query'
 import { GraphQLClient } from 'graphql-request'
 
-import { useAuthenticatedFetch } from './useAuthenticatedFetch.js'
+import { useAuthenticatedFetch } from './'
 
 /**
  * A hook for querying admin data.
